requirements.txt
gspread
oauth2client
osuapi
pprint
glob
discord
pillow